Move iconcellsignal out of icons folder
