+++ static/app/components/badge/iconCellSignal.tsx
